% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fepsfrontieR.R
\docType{package}
\name{fepsfrontieR}
\alias{fepsfrontieR}
\alias{fepsfrontieR-package}
\title{fepsfrontieR: A package for estimation of fixed-effect panel stochastic
frontier models by within-model transformation as proposed by Wang & Ho.
As proposed, Maximum Likelihood Estimation using nlminb().
Confidence Intervals are provided using the Hessian Matrix or
(time intensive) Bootstrapping to calculate Standard Error.
In addition, a Gibbs Sampler can be chosen as an alternative to MLE.}
\description{
The fepsfrontieR package provides three important functions:
sfmfep, SFM.within and SFM.generate
}
\section{sfmfep}{

The sfmfep functions ...
}

\section{SFM.within}{

The SFM.within functions ...
}

\section{SFM.generate}{

The SFM.generate functions ...
}

