% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelgdp.R
\docType{data}
\name{panelgdp}
\alias{panelgdp}
\title{Stochastic Frontier Model of the GDP of countries}
\format{A data frame with 60 rows and k+r+1 variables:
\describe{
  \item{country}{82 countries with observations each}
  \item{code}{Each country has its own code}
  \item{yr}{The year of the obseration, starting from 1 to 28}
  \item{y}{...}
  \item{k}{...}
  \item{l}{...}
  \item{h}{...}
}}
\usage{
panelgdp
}
\description{
Stochastic Frontier Model of the GDP of countries
}
\keyword{datasets}
