% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm.plot.R
\name{plot.sfmfep}
\alias{plot.sfmfep}
\title{Plot of the Inefficencys}
\usage{
\method{plot}{sfmfep}(x, ...)
}
\arguments{
\item{x}{is a data frame of the inefficencys per panel.}

\item{...}{Additional arguments to the function}
}
\description{
Plots the Inefficencys of each panel.
}
