% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm.print.R
\name{print.sfmfep}
\alias{print.sfmfep}
\title{Print Output for sfmfep}
\usage{
\method{print}{sfmfep}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{is a data frame of the inefficencys per panel.}

\item{digits}{defines the amount of digits for the values of the summary}

\item{...}{Additional arguments to the function}
}
\description{
A print of valuable variables is done when sfmfep is called.
}
