% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm_bootstrap.R
\name{SFM.bootstrap}
\alias{SFM.bootstrap}
\title{Generates B Individual Bootstrap Samples from the data and performes MLE
for each sample. Unlike i.i.d. bootstrapping, individual bootrapping samples
the rows with replacement individually for each panel instead from all samples.
In addition to the Mean and the Standard Error of the estimates,
a Confidence Interval is returned based on the quantiles of the distribution of estimates.}
\usage{
SFM.bootstrap(y, xv, z, mu, N, Time, method, R, K, B, myPar = NULL,
  lowerInt, sigmaCI)
}
\arguments{
\item{y}{is a n*t x 1 vector (response)}

\item{xv}{is a n*t x k matrix (explantatory variables)}

\item{z}{is a n*t x r matrix (inefficency determinants)}

\item{mu}{is an integer (mean of the truncated normal distribution of the inefficency)}

\item{N}{is an integer (n - panels)}

\item{Time}{is an integer (observations per panel)}

\item{method}{a required string specifying the method ("within" or "firstdiff").}

\item{R}{is an integer (# of z variables)}

\item{K}{is an integer (# of xv variables)}

\item{B}{is an integer (# of Bootstraps)}

\item{myPar}{is a vecor which has to be entered in the following order:
c(sigma_v, sigma_u, beta = c(), delta = c()). Required as starting point for the estimation.}

\item{lowerInt}{is a vector of doubles (lower bound for the estimation)}

\item{sigmaCI}{is a vector of doubles (significance of the Confidence Intervals)}
}
\value{
A B x k+r+2 matrix is returned of the estimates, the mean, standard error
     and a Confidence Interval for each estimate as a data frame.
}
\description{
Generates B Individual Bootstrap Samples from the data and performes MLE
for each sample. Unlike i.i.d. bootstrapping, individual bootrapping samples
the rows with replacement individually for each panel instead from all samples.
In addition to the Mean and the Standard Error of the estimates,
a Confidence Interval is returned based on the quantiles of the distribution of estimates.
}
