% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm.generate.R
\name{SFM.generate}
\alias{SFM.generate}
\title{Data Generation for Stochastic Frontier Models}
\usage{
SFM.generate(N, Time, beta, delta, sigma_u, sigma_v, mu = 0)
}
\arguments{
\item{N}{is an integer and specifies the amount of panels}

\item{Time}{is an integer and specifies the amount of observations per panel}

\item{beta}{is a vector of k estimates.
K explenatory varibles will be gerenated.}

\item{delta}{is a vector of r estimates.
R inefficency determinats will be gerenated.}

\item{sigma_u}{is postivie numeric and is the variation
of the stochastic inefficency}

\item{sigma_v}{is postivie numeric and is the variation
of the zero-mean random error.}

\item{mu}{is numeric and is the mean of the truncated normal distribution
of the stochastic inefficency.}
}
\value{
A data.frame( ) including x, y, z & alpha variables
}
\description{
Creates a fixed-effects panel stochastic frontier model
    Data sets created by this function can be applied for sfmfep( ) of
    this package fepsfrontieR. The model specificiation are:
    alpha_i are fixed-effect parameters drawn from a uniform distribution in [0, 1].
    x_it is drawn from a normal distribution N(alpha_i, 1).
    z_it is drawn from a standard normal distribution.
    u_i (inefficency) is drawn from a truncated normal distribution with N(mu, sigma_u).
}
\examples{
    exampleSFM <- SFM.generate (N = 20, Time = 5, beta = c(0.5, 0.2, 9, 3),
        delta = c(0.5, 0.1, 2), sigma_u = 0.2, sigma_v = 0.1)

    exampleSFM
}
