% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm_inefficency_index.R
\name{SFM.inindex}
\alias{SFM.inindex}
\title{< calculate the inefficiency of each individuum >}
\usage{
SFM.inindex(h, sigma2star, mu2star, N, Time)
}
\arguments{
\item{par}{is a vector of regression coefficients & variance parameters.
1st parameter: sigma_u, 2nd parameter: sigma_v, followed by K beta & R delta coefficients}
}
\value{
< returns the inefficiency for each individuum >
}
\description{
< calculate the inefficiency of each individuum >
}
\examples{
< an example >
}
