% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm_alpha.R
\name{SFM.alpha}
\alias{SFM.alpha}
\title{< Recovers the alpha (inteysfdysyxyssfdrcept) of each panel >}
\usage{
SFM.alpha(y, x, beta, sigma_u, sigma_v, h, epsilon, N, Time, mu = 0)
}
\arguments{
\item{par}{is a vector of regression coefficients & variance parameters.
1st parameter: sigma_u, 2nd parameter: sigma_v, followed by K beta & R delta coefficients}
}
\value{
< Describe what is returned when applying this functinon >
}
\description{
< Recovers the alpha (inteysfdysyxyssfdrcept) of each panel >
}
\examples{
< an example >
}
