% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm_alpha.R
\name{SFM.alpha}
\alias{SFM.alpha}
\title{Alpha}
\usage{
SFM.alpha(y, x, beta, sigma_u, sigma_v, h, epsilon, N, Time, mu, cumTime)
}
\arguments{
\item{y}{is a n*t x 1 vector (response)}

\item{x}{is a n*t x k matrix (explantatory variables)}

\item{N}{an optional integer specifying the total amount of panels in the data set.}

\item{Time}{an optional integer specifying the amount of observations per panel.}

\item{mu}{is the mean of a truncated normal distribution of the stochastic inefficencys.}
}
\value{
recovered values of individual fixed effects
}
\description{
Recovers the individual fixed effects by using the first-order condition for
alpha from the untransformed log-likelihood function of the model assuming
all other parameters are known. Altough, the variance of alpha is high for small
sample sized date.
}
