% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm_ci.R
\name{SFM.CI}
\alias{SFM.CI}
\title{Confidence Intervals}
\usage{
SFM.CI(estimates, hessianMatrix, alpha, N, Time, df)
}
\arguments{
\item{par}{is a vector of regression coefficients & variance parameters.
1st parameter: sigma_u, 2nd parameter: sigma_v, followed by K beta & R delta coefficients}
}
\value{
< Describe what is returned when applying this functinon >
}
\description{
Computes Confidence Intervals for the regression coefficients based on the Hessian Matrix
}
\examples{
< an example >
}
