% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm_generate.R
\name{SFM.generate}
\alias{SFM.generate}
\title{SFM.generate creates a fixed-effects panel stochastic frontier model data set
which can be used as examples for other functions of the package fepsfrontieR.}
\usage{
SFM.generate(N, Time, beta, delta, sigma_u, sigma_v, mu_u = 0)
}
\arguments{
\item{N}{is an integer and specifies the amount of panels}

\item{Time}{is an integer and specifies the amount of observations per panel}

\item{beta}{is a vector of k estimates.
K explenatory varibles will be gerenated.}

\item{delta}{is a vector of r estimates.
R inefficency determinats will be gerenated.}

\item{sigma_u}{is postivie numeric and is the variation
of the stochastic inefficency}

\item{sigma_v}{is postivie numeric and is the variation
of the zero-mean random error.}

\item{mu_u}{is numeric and is the mean of the truncated normal distribution
of the stochastic inefficency.}
}
\value{
A data.fram() including x, y, z variables
}
\description{
SFM.generate creates a fixed-effects panel stochastic frontier model data set
which can be used as examples for other functions of the package fepsfrontieR.
}
\examples{
< an example >
}
