% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm_inefficency_index.R
\name{SFM.inindex}
\alias{SFM.inindex}
\title{< What it does >}
\usage{
SFM.inindex(h, sigma2star, mu2star, N, Time)
}
\arguments{
\item{par}{is a vector of regression coefficients & variance parameters.
1st parameter: sigma_u, 2nd parameter: sigma_v, followed by K beta & R delta coefficients}
}
\value{
< Describe what is returned when applying this functinon >
}
\description{
< What it does >
}
\examples{
< an example >
}
