% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm.inefficency.R
\name{SFM.inindex}
\alias{SFM.inindex}
\title{Inefficiency Index}
\usage{
SFM.inindex(h, sigma2star, mu2star, N, Time, method, cumTime)
}
\arguments{
\item{par}{is a vector of regression coefficients & variance parameters.
1st parameter: sigma_u, 2nd parameter: sigma_v, followed by K beta & R delta coefficients}
}
\value{
returns the mean ineffieciency of each panel
}
\description{
Estimator for the inefficiency based on the conditional expectation of "u"
on "epsilon". The estimator does not require alpha and thus not suffer from the
approximation problem.
}
