% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm.inefficency.R
\name{SFM.inindex}
\alias{SFM.inindex}
\title{Inefficiency Index of the Stochastic Frontier Model}
\usage{
SFM.inindex(h, sigma2star, mu2star, N, cumTime)
}
\arguments{
\item{h}{are the values of the positive non-stochastic inefficency determinants.}

\item{sigma2star}{is a vector of variance estimations based on a chosen transformation.}

\item{mu2star}{is a vector of mean estimations based on a chosen transformation.}

\item{N}{an optional integer specifying the total amount of panels in the data set.}

\item{cumTime}{ia a vector of the cumulated times of the Time vector.
It serves as an index for computation.}
}
\value{
returns the mean ineffieciency of each panel as a Nx1 vector
}
\description{
Estimation of the inefficiency based on the conditional expectation of "u"
    on "epsilon". The estimator does not require alpha and thus not suffer from the
    approximation problem.
}
