% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm.within.R
\name{SFM.within}
\alias{SFM.within}
\title{Within Transformation of a Stochastic Frontier Model}
\usage{
SFM.within(par = c(sigma_u, sigma_v, beta = c(), delta = c()), xv, y, z,
  N = NULL, Time = NULL, cumTime, mu = 0, optim = F, K = NULL,
  R = NULL, seqN = 1:N)
}
\arguments{
\item{par}{is a vector of regression coefficients & variance parameters.
1st parameter: sigma_u, 2nd parameter: sigma_v, followed by K beta & R delta coefficients}

\item{xv}{is a n*t x k matrix (explantatory variables)}

\item{y}{is a n*t x 1 vector (response)}

\item{z}{is a n*t x r matrix (inefficency determinants)}

\item{N}{is a integer (n - panels)}

\item{Time}{is a integer (observations per panel)}

\item{cumTime}{ia a vector of the cumulated times of the Time vector.
It serves as an index for computation.}

\item{mu}{is a number (mean of the truncated normal distribution of the inefficency)}

\item{optim}{is a boolean (set F to obtain a list of model variables.
T to obtain the -sum of log.likelihood)}

\item{K}{is an integer (# of xv variables)}

\item{R}{is an integer (# of z variables)}

\item{seqN}{is a sequence from 1 to N}
}
\value{
If optim = T the log.likelihood is returned of all panels.
    If optim = F the model fit is returned including all important model variables.
}
\description{
Performs a within transformation to a Stochastic Frontier Model.
    By within-transformation, the sample mean of each panel is subtracted
    from every observation in the panel. The transformation thus removes
    the time-invariant individual effect from the model
}
