% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfmfep.R
\name{sfmfep}
\alias{sfmfep}
\title{sfmfep estimates a stochastic frontier model for fixed-effects using
within transformation}
\usage{
sfmfep(formula, data, group = NULL, N = NULL, Time = NULL, mu = 0,
  sigmaCI = 0.05, estimate = T, myPar = c(sigma_u = NULL, sigma_v =
  NULL, beta = c(NULL), delta = c(NULL)))
}
\arguments{
\item{formula}{ist etwas}

\item{par}{is a vector of regression coefficients & variance parameters.
1st parameter: sigma_u, 2nd parameter: sigma_v, followed by
K beta & R delta coefficients}
}
\value{
< Describe what is returned when applying this functinon >
}
\description{
sfmfep estimates a stochastic frontier model for fixed-effects using
within transformation
}
\examples{
< an example where a data set is created via sfm.generate>
}
