% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfmfep.R
\name{sfmfep}
\alias{sfmfep}
\title{Fixed effect stochastic Frontier Model}
\usage{
sfmfep(formula, data, panel = NULL, N = NULL, Time = NULL,
  method = "firstdiff", mu = 0, sigmaCI = 0.05, estimate = T,
  bootstrap = F, B = NULL, myPar = c(sigma_u = NULL, sigma_v = NULL,
  beta = c(NULL), delta = c(NULL)))
}
\arguments{
\item{formula}{an object of class "formula" in the form of
y ~ x1 + ... + x_k + (z1 + ... + z_r). The details of model specification are given under Details}

\item{data}{an optional data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{panel}{an optional vector specifying the panels to be used in the fitting process.}

\item{N}{an optional integer specifying the total amount of panels in the data set.
If N is entered, Time is also a required input.}

\item{Time}{an optional integer specifying the amount of observations per panel.
If Time is entered, N is also a required input.}

\item{method}{a required string specifying the method ("within" or "firstdiff").}

\item{mu}{is the mean of a truncated normal distribution of the stochastic inefficiency.}

\item{sigmaCI}{is an optional vector specifying the significance
values of the confidence intervals for the MLE estimates.}

\item{estimate}{TRUE or FALSE specifies if "myPar" is used as
starting point of the estimation, or if "myPar" is used to fit
a given stochastic frontier model.}

\item{B}{is a required input for Bootstrap = TRUE. It defines
the amount of bootstrap samples.}

\item{myPar}{is a vecor which has to be entered in the following order:
c(sigma_v, sigma_u, beta = c(), delta = c())}

\item{Bootstrap}{is an optional boolean variable. If it is set to TRUE, bootstrapping is
performed. "B" must be specified.}
}
\value{
sfmfep returns an object of class S3. The function summary( )
    can be used to obtain or print a summary of the results.
    An object of class "sfmfep" is a list containing at least the
    following components:
    TODO(): Add all output components here
}
\description{
Estimates a Stochastic Frontier Model for Fixed-Effects.
sfmfep is used to fit fixed-effect stochastic frontier models for panel data
using a specified model transformation. Bootstrapping can be performed to
calculate the standard errors instead of a numerical deriviation
via the hessian matrix.
}
\examples{
data <- sfm.pdata1

wh.1 <- sfmfep(formula = y ~ x1 + x2 + (z1), bootstrap = T,
        B = 500, method = "firstdiff", N = 100, Time = 6, data = data)
summary(wh.1)
p.gdp <- sfmfep(formula = y ~ k + l + (h), bootstrap = T, B = 10,
         method = "firstdiff", N = 82, Time = 28, data = panelgdp)
summary(p.gdp)
}
