% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm.summary.R
\name{summary.sfmfep}
\alias{summary.sfmfep}
\title{Summarizes a sfmfep fitted model}
\usage{
\method{summary}{sfmfep}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{is a data frame of the inefficencys per panel.}

\item{digits}{defines the amount of digits for the values of the summary}

\item{...}{Additional arguments to the function}
}
\description{
\code{summary.sfmfep} is the sfmfep specific method for
the generic function \code{summary} which summarize objects
returned by modelling functions.
}
